requirements.txt
numpy==1.21.0
pandas==1.3.0
scikit-learn==0.24.2
tensorflow==2.6.0
opencv-python==4.5.3.20210927
matplotlib==3.4.3
seaborn==0.11.2
jupyter==1.0.0
pytest==6.2.4
Pillow==8.4.0
ultralytics
torch
torchvision
carla